requirements.txt
streamlit>=1.40
numpy>=1.24
pandas>=2.0
pillow>=10.0
scikit-image>=0.22
scikit-learn>=1.4
scipy>=1.10
matplotlib>=3.8
joblib>=1.3
kaggle>=1.6
